pyproject.toml
[project]
name = "backend"
version = "0.3.8"
description = "API WEB HOSPITAL-SDLG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.0.0",
    "cachetools>=5.5.2",
    "click>=8.1.8",
    "cryptography>=44.0.2",
    "faker>=37.3.0",
    "fastapi[all]>=0.115.12",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "orjson>=3.10.16",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "polars>=1.32.2",
    "polars-lts-cpu>=1.33.1",
    "psycopg>=3.2.6",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.2",
    "pyjwt>=2.10.1",
    "pyrefly>=0.27.0",
    "pytesseract>=0.3.13",
    "pytest>=8.3.5",
    "pytest-rich>=0.2.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "scalar-fastapi>=1.0.3",
    "sqlmodel>=0.0.24",
    "stripe>=12.4.0",
]

[tool]

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = ['**/*venv/**\*']
